Some optimization on power calculations
Changed the calculation when having an integer exponent to calculate the
value for the power function via a loop and not the actual function
implemented in the math library.
Additionally, I added a gitignore file, to remove different files from
being tracked (like .pyc, .o, ... files).
Changed the pointers for the integration functions into smart pointers,
allowing them to be deleted automatically, when they go out of scope.