Refactored the code
Split the code base into different files, which are named according to
their classes. Also added a file specifically for the implementation of
the module for python.
Due to these changes, the setup.py file also had to be changed.
Finally, I added an exception class, which is thrown, when the user
chooses a non-existent implementation scheme.