Various fixes and Redesign
Created a wrapper class for the python interface for the GCE. This
interface allows to test the C++ code base directly instead of only
allowing to test the python interface.
Changed error handling upon return of bad values, this now throws an
error, if a NAN value is encountered.
Removed some clutter in the Dihedral entropy, also renamed a variable.