Empty list fixes (#6)
* Fixed Error that killed Python

A hugely bad bug killed the Python Kernel when giving the GCE an empty
list. This commit should fix this and also closes #3.

* Added Translator for Exceptions

As Python cannot read the exceptions from C++ and they would directly go
deep into the python kernel, I decided to translate one error (for now)
to python, this allows for testing if this actually works.

* Removed commented include

A remainder of some misguided concept of mine was removed.

* Fixed wrong error message

A wrong error message was displayed, this is fixed now. Furthermore, the
exception translator now takes a const reference.

Former-commit-id: 0240d85c7b53cd57962d53aeff38da972bb60fbc