fix type, add tests for method arg
fixed typo in docstring. Since I was already on it I also added
a function which handles the method argument. Unknown integrators
will already be caught here. also variations of upper and lower
caps will be dealt with. Therefore: Simpson, simpson or sImPsOn
will all be okay as an argument...