Added an Exception Handler for OpenMP code.
Since OpenMP is not allowed to exit via an exception, I added a class
that catches all the exceptions and then rethrows the first one, this
could be improved upon, by rethrowing all the exceptions in one
exception. However, for now this should be enough and I should not
believe that too many exceptions will occur at the same time.
HOWEVER, I do not know how to test against this behavior from the C++
side, so no new test has been added.