Handle resolution (#14)
* process user input, documentation

resolution argument is being processed separately.
Introduced the verbose argument to calculate entropy.
some doc strings.

* fix type, add tests for method arg

fixed typo in docstring. Since I was already on it I also added
a function which handles the method argument. Unknown integrators
will already be caught here. also variations of upper and lower
caps will be dealt with. Therefore: Simpson, simpson or sImPsOn
will all be okay as an argument...

* Removed calculation of base 2.

* Fixed error in the Python Interface

I forgot to remove the Python function that corresponds to the C++
function, which would create a GCE from only a list. This is not
possible anymore.

Co-authored-by: Patrick Quoika <Patrick.Quoika@uibk.ac.at>
Former-commit-id: 7b26dedfe12f0d8c8e332a9b9efbe2a758886c41