Added an Exception Handler for OpenMP code. (#16)
* Added an Exception Handler for OpenMP code.

Since OpenMP is not allowed to exit via an exception, I added a class
that catches all the exceptions and then rethrows the first one, this
could be improved upon, by rethrowing all the exceptions in one
exception. However, for now this should be enough and I should not
believe that too many exceptions will occur at the same time.
HOWEVER, I do not know how to test against this behavior from the C++
side, so no new test has been added.

* Added the OMPExceptionHandler to DihedralEntropy

Before this was not done, which did not solve the issue, as both
DihedralEntropy and Gce have their own integration functions. This might
get fixed in another commit. As this is just duplicated code.
The exception will be translated to Python in a future commit.

* Added pragma once

Compilation did not work, because I forgot to define pragma once at the
top of the header file. this is fixed now.