Baby steps with Cython
I added the first step in moving from boost to Cython.
I moved some functionallity that was done in boost to Cython. Cython is more strongly connected to Python.
On Windows it does not yet work, since the compilation arguments must be given differently.
Specifically, on Windows, with the Windows compiler, the '-' is replaced by a '/'.