Finalized migration to Cython
Unnecessary files have been deleted, so all remnants of boost are gone.
Furhtermore, a small bug in the C++ implementation was fixed, where the
FFT was calculated over 1 value more than was allocated. Weirdly enough
this did not lead to a segmentation fault as one might expect, but
simply to wrong results.
The get_grid method now actually returns the centers. I also updated
the names of the Python functions to follow Python convention (sorry
Patrick, this will probably break your code).
The Tests for the GCE were updated.