Cython (#20)
* Started changes to Cython

Removed boost necessary files and started preparing for Cython.

* Baby steps with Cython

I added the first step in moving from boost to Cython.
I moved some functionallity that was done in boost to Cython. Cython is more strongly connected to Python.
On Windows it does not yet work, since the compilation arguments must be given differently.
Specifically, on Windows, with the Windows compiler, the '-' is replaced by a '/'.

* Fixed small include error

In Exceptions.h the correct file was not included, but now everything should be fine.

* Finished Cython interface

The Cython interface can now return the most important things, i.e., the
estimated PDF, the grid and the bandwidth of the kernel.
The integration is still misleading, as this actually calculates the
entropy. However, this will be fixed in a future commit.

* Fix in github action

* Another Fix in github Action

* change submodule name and setup

intermediate commit to avoid clashes of kde front and backend
class.

* Introduce kde frontend (#21)

* scratch kde class for python frontend

direction is already visible, but I need to rebase on the new
c++ backend with cython interface instead of boost

* adapt python frontend to cython backend

properties of the kde class are now properly filled and initialized and
the python backend functions are already being used. dihedrals still
broken... what is up with the bin-centering/ pdf

* removed liboost from pyaction

Co-authored-by: Patrick Quoika <Patrick.Quoika@uibk.ac.at>

* Fixed uninitialized error.

* Finalized migration to Cython

Unnecessary files have been deleted, so all remnants of boost are gone.
Furhtermore, a small bug in the C++ implementation was fixed, where the
FFT was calculated over 1 value more than was allocated. Weirdly enough
this did not lead to a segmentation fault as one might expect, but
simply to wrong results.
The get_grid method now actually returns the centers. I also updated
the names of the Python functions to follow Python convention (sorry
Patrick, this will probably break your code).
The Tests for the GCE were updated.

* apply function renaming

Co-authored-by: Patrick Quoika <Patrick.Quoika@uibk.ac.at>
Co-authored-by: Patrick Quoika <i.ah.esel@web.de>
Former-commit-id: 6b7b1f4d9062096a844ece601b8e2e8202a8db55