first scratch
still a lot of legacy code and so on. but the scratch is already quite
visible. still need to change the setup for the reweighting module also
I am dependent on a new cython function that for reweighted histograms
for the kde. apart from that docstrings are still old and so on... not
  too much work anymore though, I'd say...


Former-commit-id: 6ec5f000b6c85e85d376855afe23953e02264fc8