continue design of dihedral entropy class
Still missing the counterpart from the cython kernel for dihedrals with
weights. So far the pipeline only accepts one data set. Would probably
be a good idea to take out subroutines from the calculate functions, for
them to be called in a loop of a collection of data sets. Didnt test
anything, yet...


Former-commit-id: f940ba2f5e82fef0429047d498e00542080d0232