change dihedrals from deg to rad
We calculate in rad now. The users input will be sanity checked for the
right dihedral units and then be transformed. a warning is thrown. in
tour course I made constants an internal package (just more handy, but
also it makes sense, I think) Problem now is the plotting in the
examples: should I transfer back the pdf to input units?