Fix bug in normalizer
Due to the way std::accumulate decides which version is used (the initial value is the deciding factor), the sum was performed on integers, this leads to 0 values, when the parts of the sum are smaller 1. This should address and close #56 , and close #57.
Tests against this behavior should be added. Do not merge until I added these tests!